- Upgrade

+++ src/Console/InstallCommand.php
@@ -3,8 +3,11 @@
 namespace Asciisd\Knet\Console;
 
 use Illuminate\Console\Command;
+use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Str;
+use Symfony\Component\Console\Attribute\AsCommand;
 
+#[AsCommand(name: 'knet:install')]
 class InstallCommand extends Command
 {
     /**
@@ -26,20 +29,19 @@ class InstallCommand extends Command
      *
      * @return void
      */
-    public function handle()
+    public function handle(): void
     {
-        $this->comment('Publishing Knet Service Provider...');
-        $this->callSilent('vendor:publish', ['--tag' => 'knet-provider']);
+        $this->components->info('Installing KNet resources.');
 
-        $this->comment('Publishing Knet Views...');
-        $this->callSilent('vendor:publish', ['--tag' => 'knet-views']);
-
-        $this->comment('Publishing Knet Configuration...');
-        $this->callSilent('vendor:publish', ['--tag' => 'knet-config']);
+        collect([
+            'Service Provider' => fn () => $this->callSilent('vendor:publish', ['--tag' => 'knet-provider']) == 0,
+            'Views' => fn () => $this->callSilent('vendor:publish', ['--tag' => 'knet-views']) == 0,
+            'Configurations' => fn () => $this->callSilent('vendor:publish', ['--tag' => 'knet-config']) == 0,
+        ])->each(fn ($task, $description) => $this->components->task($description, $task));
 
         $this->registerKnetServiceProvider();
 
-        $this->info('Knet scaffolding installed successfully.');
+        $this->components->info('Knet scaffolding installed successfully.');
     }
 
     /**
@@ -47,21 +49,25 @@ class InstallCommand extends Command
      *
      * @return void
      */
-    protected function registerKnetServiceProvider()
+    protected function registerKnetServiceProvider(): void
     {
         $namespace = Str::replaceLast('\\', '', $this->laravel->getNamespace());
 
-        $appConfig = file_get_contents(config_path('app.php'));
+        if (file_exists($this->laravel->bootstrapPath('providers.php'))) {
+            ServiceProvider::addProviderToBootstrapFile("{$namespace}\\Providers\\KnetServiceProvider");
+        } else {
+            $appConfig = file_get_contents(config_path('app.php'));
 
-        if (Str::contains($appConfig, $namespace.'\\Providers\\KnetServiceProvider::class')) {
-            return;
+            if (Str::contains($appConfig, $namespace.'\\Providers\\KnetServiceProvider::class')) {
+                return;
+            }
+
+            file_put_contents(config_path('app.php'), str_replace(
+                "{$namespace}\\Providers\EventServiceProvider::class,".PHP_EOL,
+                "{$namespace}\\Providers\EventServiceProvider::class,".PHP_EOL."        {$namespace}\Providers\KnetServiceProvider::class,".PHP_EOL,
+                $appConfig
+            ));
         }
-
-        file_put_contents(config_path('app.php'), str_replace(
-            "{$namespace}\\Providers\EventServiceProvider::class,".PHP_EOL,
-            "{$namespace}\\Providers\EventServiceProvider::class,".PHP_EOL."        {$namespace}\Providers\KnetServiceProvider::class,".PHP_EOL,
-            $appConfig
-        ));
 
         file_put_contents(app_path('Providers/KnetServiceProvider.php'), str_replace(
             "namespace App\Providers;",
